Moving FakeBigQueryServices to published artifacts
