+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryAvroUtils.java
@@ -57,7 +57,7 @@ import org.joda.time.format.DateTimeFormatter;
  * <p>These utilities are based on the <a href="https://avro.apache.org/docs/1.8.1/spec.html">Avro
  * 1.8.1</a> specification.
  */
-class BigQueryAvroUtils {
+public class BigQueryAvroUtils {
 
   /**
    * Defines the valid mapping between BigQuery types and native Avro types.
@@ -353,7 +353,7 @@ class BigQueryAvroUtils {
         unionTypes.get(1).getType(), unionTypes.get(1).getLogicalType(), fieldSchema, v);
   }
 
-  static Schema toGenericAvroSchema(String schemaName, List<TableFieldSchema> fieldSchemas) {
+  public static Schema toGenericAvroSchema(String schemaName, List<TableFieldSchema> fieldSchemas) {
     List<Field> avroFields = new ArrayList<>();
     for (TableFieldSchema bigQueryField : fieldSchemas) {
       avroFields.add(convertField(bigQueryField));
